Move `DecCBOR Annotator` instances back to the main libraries
from testlib
