+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Scripts.hs
@@ -65,6 +65,7 @@ import Cardano.Ledger.Alonzo.Era (AlonzoEra)
 import Cardano.Ledger.Alonzo.TxCert ()
 import Cardano.Ledger.BaseTypes (ProtVer (..), kindObject)
 import Cardano.Ledger.Binary (
+  Annotator,
   DecCBOR (decCBOR),
   DecCBORGroup (..),
   Decoder,
@@ -76,13 +77,14 @@ import Cardano.Ledger.Binary (
   encodeWord8,
  )
 import Cardano.Ledger.Binary.Coders (
-  Decode (D, From, Invalid, SumD, Summands),
+  Decode (Ann, D, From, Invalid, SumD, Summands),
   Encode (Sum, To),
   Wrapped (..),
   decode,
   encode,
   (!>),
   (<!),
+  (<*!),
  )
 import Cardano.Ledger.Binary.Plain (serializeAsHexText)
 import Cardano.Ledger.Core
@@ -609,6 +611,22 @@ instance AlonzoEraScript era => DecCBOR (AlonzoScript era) where
         n -> Invalid n
       decodePlutus slang =
         SumD PlutusScript <! D (decodePlutusScript slang)
+
+instance AlonzoEraScript era => DecCBOR (Annotator (AlonzoScript era)) where
+  decCBOR = decode (Summands "AlonzoScript" decodeScript)
+    where
+      decodeAnnPlutus slang =
+        Ann (SumD PlutusScript) <*! Ann (D (decodePlutusScript slang))
+      {-# INLINE decodeAnnPlutus #-}
+      decodeScript :: Word -> Decode 'Open (Annotator (AlonzoScript era))
+      decodeScript = \case
+        0 -> Ann (SumD TimelockScript) <*! From
+        1 -> decodeAnnPlutus SPlutusV1
+        2 -> decodeAnnPlutus SPlutusV2
+        3 -> decodeAnnPlutus SPlutusV3
+        n -> Invalid n
+      {-# INLINE decodeScript #-}
+  {-# INLINE decCBOR #-}
 
 -- | Verify that every `Script` represents a valid script. Force native scripts to Normal
 -- Form, to ensure that there are no bottoms and deserialize `Plutus` scripts into a
