+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Tx.hs
@@ -96,10 +96,12 @@ import Cardano.Ledger.Alonzo.TxWits (
   unTxDatsL,
  )
 import Cardano.Ledger.Binary (
+  Annotator,
   DecCBOR (..),
   EncCBOR (encCBOR),
   Encoding,
   ToCBOR (..),
+  decodeNullMaybe,
   decodeNullStrictMaybe,
   encodeListLen,
   encodeNullMaybe,
@@ -121,6 +123,7 @@ import Data.Aeson (ToJSON (..))
 import qualified Data.ByteString.Lazy as LBS
 import Data.Maybe.Strict (
   StrictMaybe (..),
+  maybeToStrictMaybe,
   strictMaybeToMaybe,
  )
 import Data.Set (Set)
@@ -419,6 +422,29 @@ instance
         <! From
         <! From
         <! D (decodeNullStrictMaybe decCBOR)
+  {-# INLINE decCBOR #-}
+
+instance
+  ( Typeable era
+  , Typeable (TxBody era)
+  , Typeable (TxWits era)
+  , Typeable (TxAuxData era)
+  , DecCBOR (Annotator (TxBody era))
+  , DecCBOR (Annotator (TxWits era))
+  , DecCBOR (Annotator (TxAuxData era))
+  ) =>
+  DecCBOR (Annotator (AlonzoTx era))
+  where
+  decCBOR =
+    decode $
+      Ann (RecD AlonzoTx)
+        <*! From
+        <*! From
+        <*! Ann From
+        <*! D
+          ( sequence . maybeToStrictMaybe
+              <$> decodeNullMaybe decCBOR
+          )
   {-# INLINE decCBOR #-}
 
 alonzoEqTxRaw :: AlonzoEraTx era => Tx era -> Tx era -> Bool
