+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxBody.hs
@@ -96,6 +96,7 @@ import Cardano.Ledger.BaseTypes (
   StrictMaybe (..),
  )
 import Cardano.Ledger.Binary (
+  Annotator,
   DecCBOR (..),
   EncCBOR (..),
   ToCBOR (..),
@@ -114,6 +115,7 @@ import Cardano.Ledger.Mary.Value (
  )
 import Cardano.Ledger.MemoBytes (
   EqRaw,
+  Mem,
   MemoBytes,
   MemoHashIndex,
   Memoized (..),
@@ -364,6 +366,8 @@ deriving instance Show (TxBody AlonzoEra)
 
 deriving newtype instance DecCBOR (TxBody AlonzoEra)
 
+deriving via Mem AlonzoTxBodyRaw instance DecCBOR (Annotator (TxBody AlonzoEra))
+
 pattern AlonzoTxBody ::
   Set TxIn ->
   Set TxIn ->
@@ -590,6 +594,9 @@ instance DecCBOR AlonzoTxBodyRaw where
         , (1, "outputs")
         , (2, "fee")
         ]
+
+instance DecCBOR (Annotator AlonzoTxBodyRaw) where
+  decCBOR = pure <$> decCBOR
 
 emptyAlonzoTxBodyRaw :: AlonzoTxBodyRaw
 emptyAlonzoTxBodyRaw =
