+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/TxWits.hs
@@ -75,6 +75,7 @@ import Cardano.Ledger.Alonzo.Scripts (
   toPlutusSLanguage,
  )
 import Cardano.Ledger.Binary (
+  Annotator,
   DecCBOR (..),
   DecCBORGroup (..),
   Decoder,
@@ -104,6 +105,7 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.Keys (BootstrapWitness, WitVKey)
 import Cardano.Ledger.MemoBytes (
   EqRaw (..),
+  Mem,
   MemoBytes,
   Memoized (..),
   eqRawType,
@@ -123,6 +125,7 @@ import Cardano.Ledger.Plutus.Language (
  )
 import Cardano.Ledger.Shelley.TxWits (
   ShelleyTxWits (..),
+  mapTraverseableDecoderA,
   shelleyEqTxWitsRaw,
  )
 import Control.DeepSeq (NFData)
@@ -179,6 +182,11 @@ instance Memoized (Redeemers era) where
 -- this is how we ensure that they are not manipulated.
 newtype Redeemers era = MkRedeemers (MemoBytes (RedeemersRaw era))
   deriving newtype (Generic, ToCBOR, SafeToHash, DecCBOR)
+
+deriving via
+  Mem (RedeemersRaw era)
+  instance
+    AlonzoEraScript era => DecCBOR (Annotator (Redeemers era))
 
 deriving newtype instance AlonzoEraScript era => Eq (Redeemers era)
 
@@ -362,6 +370,18 @@ instance Era era => DecCBOR (TxDatsRaw era) where
       (TxDatsRaw . Map.fromElems hashData <$> decodeList decCBOR)
   {-# INLINE decCBOR #-}
 
+instance Era era => DecCBOR (Annotator (TxDatsRaw era)) where
+  decCBOR =
+    ifDecoderVersionAtLeast
+      (natVersion @9)
+      ( allowTag setTag
+          >> mapTraverseableDecoderA
+            (decodeNonEmptyList decCBOR)
+            (TxDatsRaw . Map.fromElems hashData . NE.toList)
+      )
+      (mapTraverseableDecoderA (decodeList decCBOR) (TxDatsRaw . Map.fromElems hashData))
+  {-# INLINE decCBOR #-}
+
 -- | Note that 'TxDats' are based on 'MemoBytes' since we must preserve
 -- the original bytes for the 'Cardano.Ledger.Alonzo.Tx.ScriptIntegrity'.
 -- Since the 'TxDats' exist outside of the transaction body,
@@ -372,6 +392,11 @@ newtype TxDats era = MkTxDats (MemoBytes (TxDatsRaw era))
 
 instance Memoized (TxDats era) where
   type RawType (TxDats era) = TxDatsRaw era
+
+deriving via
+  Mem (TxDatsRaw era)
+  instance
+    Era era => DecCBOR (Annotator (TxDats era))
 
 deriving instance Show (TxDats era)
 
@@ -628,6 +653,51 @@ instance AlonzoEraScript era => DecCBOR (RedeemersRaw era) where
       {-# INLINE decodeElement #-}
   {-# INLINE decCBOR #-}
 
+instance AlonzoEraScript era => DecCBOR (Annotator (RedeemersRaw era)) where
+  decCBOR = do
+    ifDecoderVersionAtLeast
+      (natVersion @9)
+      ( peekTokenType >>= \case
+          TypeMapLenIndef -> decodeMapRedeemers
+          TypeMapLen -> decodeMapRedeemers
+          _ -> decodeListRedeemers
+      )
+      ( mapTraverseableDecoderA
+          (decodeList decodeAnnElement)
+          (RedeemersRaw . Map.fromList)
+      )
+    where
+      decodeRedeemersWith nonEmptyDecoder =
+        mapTraverseableDecoderA
+          nonEmptyDecoder
+          (RedeemersRaw . Map.fromList . NE.toList)
+      decodeMapRedeemers = decodeRedeemersWith $ do
+        (_, xs) <- decodeListLikeWithCount decodeMapLenOrIndef (:) $ \_ -> do
+          ptr <- decCBOR
+          (annData, exUnits) <- decCBOR
+          pure $ (\d -> (ptr, (d, exUnits))) <$> annData
+        case NE.nonEmpty xs of
+          Nothing -> fail "Expected redeemers map to be non-empty"
+          Just neList -> pure $ NE.reverse neList
+      decodeListRedeemers =
+        decodeRedeemersWith (decodeNonEmptyList decodeAnnElement)
+      decodeAnnElement ::
+        forall s. Decoder s (Annotator (PlutusPurpose AsIx era, (Data era, ExUnits)))
+      decodeAnnElement = do
+        (rdmrPtr, dat, ex) <- decodeElement
+        let f x y z = (x, (y, z))
+        pure $ f rdmrPtr <$> dat <*> pure ex
+      {-# INLINE decodeAnnElement #-}
+      decodeElement ::
+        forall s. Decoder s (PlutusPurpose AsIx era, Annotator (Data era), ExUnits)
+      decodeElement = do
+        decodeRecordNamed
+          "Redeemer"
+          (\(rdmrPtr, _, _) -> fromIntegral (listLen rdmrPtr) + 2)
+          $ (,,) <$> decCBORGroup <*> decCBOR <*> decCBOR
+      {-# INLINE decodeElement #-}
+  {-# INLINE decCBOR #-}
+
 -- | Encodes memoized bytes created upon construction.
 instance AlonzoEraScript era => EncCBOR (Redeemers era)
 
@@ -686,7 +756,78 @@ instance AlonzoEraScript era => DecCBOR (AlonzoTxWitsRaw era) where
           {-# INLINE pairDecoder #-}
       {-# INLINE nativeScriptsDecoder #-}
 
+instance
+  (AlonzoEraScript era, DecCBOR (Annotator (NativeScript era))) =>
+  DecCBOR (Annotator (AlonzoTxWitsRaw era))
+  where
+  decCBOR =
+    decode $
+      SparseKeyed
+        "AlonzoTxWits"
+        (pure emptyTxWitsRaw)
+        txWitnessField
+        []
+    where
+      txWitnessField :: Word -> Field (Annotator (AlonzoTxWitsRaw era))
+      txWitnessField 0 =
+        fieldAA
+          (\x wits -> wits {atwrAddrTxWits = x})
+          ( D $
+              ifDecoderVersionAtLeast
+                (natVersion @9)
+                ( allowTag setTag
+                    >> mapTraverseableDecoderA (decodeNonEmptyList decCBOR) (Set.fromList . NE.toList)
+                )
+                (mapTraverseableDecoderA (decodeList decCBOR) Set.fromList)
+          )
+      txWitnessField 1 =
+        fieldAA
+          addScriptsTxWitsRaw
+          (D nativeScriptsDecoder)
+      txWitnessField 2 =
+        fieldAA
+          (\x wits -> wits {atwrBootAddrTxWits = x})
+          ( D $
+              ifDecoderVersionAtLeast
+                (natVersion @9)
+                ( allowTag setTag
+                    >> mapTraverseableDecoderA (decodeNonEmptyList decCBOR) (Set.fromList . NE.toList)
+                )
+                (mapTraverseableDecoderA (decodeList decCBOR) Set.fromList)
+          )
+      txWitnessField 3 = fieldA addScriptsTxWitsRaw (decodeAlonzoPlutusScript SPlutusV1)
+      txWitnessField 4 =
+        fieldAA
+          (\x wits -> wits {atwrDatsTxWits = x})
+          From
+      txWitnessField 5 = fieldAA (\x wits -> wits {atwrRdmrsTxWits = x}) From
+      txWitnessField 6 = fieldA addScriptsTxWitsRaw (decodeAlonzoPlutusScript SPlutusV2)
+      txWitnessField 7 = fieldA addScriptsTxWitsRaw (decodeAlonzoPlutusScript SPlutusV3)
+      txWitnessField n = invalidField n
+      {-# INLINE txWitnessField #-}
+
+      nativeScriptsDecoder :: Decoder s (Annotator (Map ScriptHash (Script era)))
+      nativeScriptsDecoder =
+        ifDecoderVersionAtLeast
+          (natVersion @9)
+          ( allowTag setTag
+              >> mapTraverseableDecoderA (decodeNonEmptyList pairDecoder) (Map.fromList . NE.toList)
+          )
+          (mapTraverseableDecoderA (decodeList pairDecoder) Map.fromList)
+        where
+          pairDecoder :: Decoder s (Annotator (ScriptHash, Script era))
+          pairDecoder = fmap (asHashedScriptPair . fromNativeScript) <$> decCBOR
+  {-# INLINE decCBOR #-}
+
 deriving newtype instance AlonzoEraScript era => DecCBOR (AlonzoTxWits era)
+
+deriving via
+  Mem (AlonzoTxWitsRaw era)
+  instance
+    ( AlonzoEraScript era
+    , DecCBOR (Annotator (NativeScript era))
+    ) =>
+    DecCBOR (Annotator (AlonzoTxWits era))
 
 addScriptsTxWitsRaw ::
   Map ScriptHash (Script era) ->
