+++ eras/mary/impl/testlib/Test/Cardano/Ledger/Mary/Binary/Annotator.hs
@@ -3,14 +3,4 @@
 {-# LANGUAGE StandaloneDeriving #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Test.Cardano.Ledger.Mary.Binary.Annotator (
-  module Test.Cardano.Ledger.Allegra.Binary.Annotator,
-) where
-
-import Cardano.Ledger.Binary
-import Cardano.Ledger.Mary (MaryEra)
-import Cardano.Ledger.Mary.Core
-import Cardano.Ledger.Mary.TxBody
-import Test.Cardano.Ledger.Allegra.Binary.Annotator
-
-deriving via Mem MaryTxBodyRaw instance DecCBOR (Annotator (TxBody MaryEra))
+module Test.Cardano.Ledger.Mary.Binary.Annotator where
