+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Scripts.hs
@@ -34,6 +34,7 @@ module Cardano.Ledger.Shelley.Scripts (
 ) where
 
 import Cardano.Ledger.Binary (
+  Annotator,
   DecCBOR (decCBOR),
   EncCBOR (..),
   ToCBOR,
@@ -48,6 +49,7 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.Keys.WitVKey (witVKeyHash)
 import Cardano.Ledger.MemoBytes (
   EqRaw (..),
+  Mem,
   MemoBytes,
   Memoized (..),
   decodeMemoized,
@@ -113,6 +115,11 @@ newtype MultiSig era = MkMultiSig (MemoBytes (MultiSigRaw era))
 
 instance Memoized (MultiSig era) where
   type RawType (MultiSig era) = MultiSigRaw era
+
+deriving via
+  Mem (MultiSigRaw era)
+  instance
+    Era era => DecCBOR (Annotator (MultiSig era))
 
 -- | Magic number "memorialized" in the ValidateScript class under the method:
 --   scriptPrefixTag:: Core.Script era -> Bs.ByteString, for the Shelley Era.
@@ -202,6 +209,22 @@ instance Era era => DecCBOR (MultiSigRaw era) where
       1 -> (,) 2 . MultiSigAllOf <$> decCBOR
       2 -> (,) 2 . MultiSigAnyOf <$> decCBOR
       3 -> (,) 3 <$> (MultiSigMOf <$> decCBOR <*> decCBOR)
+      k -> invalidKey k
+
+instance Era era => DecCBOR (Annotator (MultiSigRaw era)) where
+  decCBOR = decodeRecordSum "MultiSig" $
+    \case
+      0 -> (,) 2 . pure . MultiSigSignature . KeyHash <$> decCBOR
+      1 -> do
+        multiSigs <- sequence <$> decCBOR
+        pure (2, MultiSigAllOf <$> multiSigs)
+      2 -> do
+        multiSigs <- sequence <$> decCBOR
+        pure (2, MultiSigAnyOf <$> multiSigs)
+      3 -> do
+        m <- decCBOR
+        multiSigs <- sequence <$> decCBOR
+        pure (3, MultiSigMOf m <$> multiSigs)
       k -> invalidKey k
 
 -- | Check the equality of two underlying types, while ignoring their binary
