+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxWits.hs
@@ -32,10 +32,13 @@ module Cardano.Ledger.Shelley.TxWits (
   bootAddrShelleyTxWitsL,
   addrWits',
   shelleyEqTxWitsRaw,
+  mapTraverseableDecoderA,
 ) where
 
 import Cardano.Ledger.Binary (
+  Annotator,
   DecCBOR (decCBOR),
+  Decoder,
   EncCBOR (encCBOR),
   decodeList,
  )
@@ -52,6 +55,7 @@ import Cardano.Ledger.Hashes (SafeToHash (..))
 import Cardano.Ledger.Keys (BootstrapWitness, KeyRole (Witness), WitVKey (..))
 import Cardano.Ledger.MemoBytes (
   EqRaw (..),
+  Mem,
   MemoBytes,
   Memoized (..),
   decodeMemoized,
@@ -69,6 +73,7 @@ import qualified Data.Map.Strict as Map
 import qualified Data.MapExtras as Map (fromElems)
 import Data.Set (Set)
 import qualified Data.Set as Set
+import Data.Void (Void)
 import GHC.Generics (Generic)
 import GHC.Records ()
 import Lens.Micro (Lens', (^.))
@@ -105,6 +110,14 @@ newtype ShelleyTxWits era = MkShelleyTxWits (MemoBytes (ShelleyTxWitsRaw era))
 
 instance Memoized (ShelleyTxWits era) where
   type RawType (ShelleyTxWits era) = ShelleyTxWitsRaw era
+
+deriving via
+  Mem (ShelleyTxWitsRaw era)
+  instance
+    ( EraScript era
+    , DecCBOR (Annotator (Script era))
+    ) =>
+    DecCBOR (Annotator (ShelleyTxWits era))
 
 deriving newtype instance EraScript era => Eq (ShelleyTxWits era)
 
@@ -232,3 +245,44 @@ instance EraScript era => DecCBOR (ShelleyTxWitsRaw era) where
           (D $ Map.fromElems (hashScript @era) <$> decodeList decCBOR)
       witField 2 = field (\x wits -> wits {stwrBootAddrTxWits = x}) From
       witField n = invalidField n
+
+instance
+  ( EraScript era
+  , DecCBOR (Annotator (Script era))
+  ) =>
+  DecCBOR (Annotator (ShelleyTxWitsRaw era))
+  where
+  decCBOR =
+    decode $
+      SparseKeyed
+        "ShelleyTxWitsRaw"
+        (pure emptyWitnessSet)
+        witField
+        []
+    where
+      emptyWitnessSet = ShelleyTxWitsRaw mempty mempty mempty
+      witField :: Word -> Field (Annotator (ShelleyTxWitsRaw era))
+      witField 0 =
+        fieldAA
+          (\x wits -> wits {stwrAddrTxWits = x})
+          (D $ mapTraverseableDecoderA (decodeList decCBOR) Set.fromList)
+      witField 1 =
+        fieldAA
+          (\x wits -> wits {stwrScriptTxWits = x})
+          ( D $
+              mapTraverseableDecoderA
+                (decodeList decCBOR)
+                (Map.fromElems (hashScript @era))
+          )
+      witField 2 =
+        fieldAA
+          (\x wits -> wits {stwrBootAddrTxWits = x})
+          (D $ mapTraverseableDecoderA (decodeList decCBOR) Set.fromList)
+      witField n = fieldAA (\(_ :: Void) wits -> wits) (Invalid n)
+
+mapTraverseableDecoderA ::
+  Traversable f =>
+  Decoder s (f (Annotator a)) ->
+  (f a -> m b) ->
+  Decoder s (Annotator (m b))
+mapTraverseableDecoderA decList transformList = fmap transformList . sequence <$> decList
