+++ libs/cardano-ledger-binary/cardano-ledger-binary.cabal
@@ -132,14 +132,12 @@ library testlib
     hspec-core,
     iproute,
     mtl,
-    plutus-ledger-api,
     pretty-show,
     prettyprinter,
     prettyprinter-ansi-terminal,
     primitive,
     quickcheck-instances >=0.3.32,
     random >=1.2,
-    serialise,
     tasty-hunit,
     text,
     tree-diff,
