+++ libs/cardano-ledger-core/src/Cardano/Ledger/Keys/Bootstrap.hs
@@ -28,7 +28,7 @@ import qualified Cardano.Crypto.DSIGN.Class as C
 import qualified Cardano.Crypto.Hash as Hash
 import qualified Cardano.Crypto.Signing as Byron
 import qualified Cardano.Crypto.Wallet as WC
-import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
+import Cardano.Ledger.Binary (Annotator, DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Plain (
   FromCBOR (..),
   ToCBOR (..),
@@ -84,10 +84,13 @@ instance EncCBOR BootstrapWitness
 
 instance FromCBOR BootstrapWitness where
   fromCBOR =
-    decodeRecordNamed "BootstrapWitnessRaw" (const 4) $
+    decodeRecordNamed "BootstrapWitness" (const 4) $
       BootstrapWitness <$> fromCBOR <*> C.decodeSignedDSIGN <*> fromCBOR <*> fromCBOR
 
 instance DecCBOR BootstrapWitness
+
+instance DecCBOR (Annotator BootstrapWitness) where
+  decCBOR = pure <$> decCBOR
 
 instance Ord BootstrapWitness where
   compare = comparing bootstrapWitKeyHash
